Tests refactor
Factored out the `core` tests -- the file was getting too big.
Also created folders for other modules for the future.
