+++ tests/errors/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the errors module."""
