--- tests/test_core.py
@@ -1,1231 +1,0 @@
-"""Tests for the core Ether model."""
-
-import json
-
-import pytest
-from pydantic import BaseModel, ValidationError
-
-from ether.attachment import Attachment
-from ether.core import Ether, _adapter_registry, _spec_registry
-from ether.errors import ConversionError, RegistrationError
-
-
-class TestEtherCreation:
-    """Test Ether model creation and validation."""
-
-    def test_create_ether_with_empty_payload_and_metadata(self) -> None:
-        """Test creating an Ether with empty payload and metadata."""
-        ether = Ether(
-            kind="embedding",
-            schema_version=1,
-            payload={},
-            metadata={},
-        )
-
-        assert ether.kind == "embedding"
-        assert ether.schema_version == 1
-        assert ether.payload == {}
-        assert ether.metadata == {}
-        assert ether.extra_fields == {}
-        assert ether.attachments == []
-
-    def test_create_ether_with_default_schema_version(self) -> None:
-        """Test creating an Ether with default schema_version."""
-        ether = Ether(
-            kind="tokens",
-            payload={"ids": [1, 2, 3]},
-            metadata={"vocab": "bert-base"},
-        )
-
-        assert ether.kind == "tokens"
-        assert ether.schema_version == 1  # default value
-        assert ether.payload == {"ids": [1, 2, 3]}
-        assert ether.metadata == {"vocab": "bert-base"}
-
-    def test_create_ether_with_attachments(self) -> None:
-        """Test creating an Ether with attachments."""
-        attachment = Attachment(
-            id="emb-0",
-            uri="shm://embeddings/12345",
-            media_type="application/x-raw-tensor",
-            codec="RAW_F32",
-            shape=[768],
-            dtype="float32",
-        )
-
-        ether = Ether(
-            kind="embedding",
-            schema_version=1,
-            payload={"dim": 768},
-            metadata={"source": "bert-base"},
-            attachments=[attachment],
-        )
-
-        assert ether.kind == "embedding"
-        assert len(ether.attachments) == 1
-        assert ether.attachments[0].id == "emb-0"
-        assert ether.attachments[0].codec == "RAW_F32"
-
-    def test_create_ether_with_extra_fields(self) -> None:
-        """Test creating an Ether with extra fields."""
-        ether = Ether(
-            kind="text",
-            payload={"text": "Hello, world!"},
-            metadata={"lang": "en"},
-            extra_fields={"note": "test message", "priority": "high"},
-        )
-
-        assert ether.extra_fields == {"note": "test message", "priority": "high"}
-
-    def test_create_ether_with_nested_payload_and_metadata(self) -> None:
-        """Test creating an Ether with nested payload and metadata."""
-        ether = Ether(
-            kind="embedding",
-            payload={"vec": {"values": [1.0, 2.0, 3.0], "dim": 3}},
-            metadata={"model": {"name": "bert-base", "version": "1.0"}},
-        )
-
-        assert ether.payload == {"vec": {"values": [1.0, 2.0, 3.0], "dim": 3}}
-        assert ether.metadata == {"model": {"name": "bert-base", "version": "1.0"}}
-
-
-class TestEtherValidation:
-    """Test Ether model validation and error cases."""
-
-    def test_validation_error_empty_kind(self) -> None:
-        """Test that empty kind is allowed for models without kind specification."""
-        # Empty kind is now allowed for models that don't specify a kind
-        ether = Ether(
-            kind="",
-            payload={},
-            metadata={},
-        )
-        assert ether.kind == ""
-
-    def test_validation_error_negative_schema_version(self) -> None:
-        """Test that negative schema_version raises an error."""
-        with pytest.raises(ValidationError):
-            Ether(
-                kind="embedding",
-                schema_version=0,
-                payload={},
-                metadata={},
-            )
-
-    def test_validation_error_zero_schema_version(self) -> None:
-        """Test that zero schema_version raises an error."""
-        with pytest.raises(ValidationError):
-            Ether(
-                kind="embedding",
-                schema_version=0,
-                payload={},
-                metadata={},
-            )
-
-    def test_validation_error_missing_required_fields(self) -> None:
-        """Test that missing required fields raises ValidationError."""
-        with pytest.raises(ValidationError):
-            Ether(
-                kind="embedding",
-                # missing payload and metadata
-            )
-
-    def test_validation_error_invalid_kind_type(self) -> None:
-        """Test that invalid kind type raises ValidationError."""
-        with pytest.raises(ValidationError):
-            Ether(
-                kind=123,  # should be str
-                payload={},
-                metadata={},
-            )
-
-    def test_validation_error_invalid_schema_version_type(self) -> None:
-        """Test that invalid schema_version type raises ValidationError."""
-        with pytest.raises(ValidationError):
-            Ether(
-                kind="embedding",
-                schema_version="invalid",  # should be int
-                payload={},
-                metadata={},
-            )
-
-
-class TestEtherSerialization:
-    """Test Ether model serialization and deserialization."""
-
-    def test_model_dump_round_trip(self) -> None:
-        """Test that model_dump() and model_validate() work correctly."""
-        original = Ether(
-            kind="embedding",
-            schema_version=2,
-            payload={"dim": 768},
-            metadata={"source": "bert-base"},
-            extra_fields={"note": "test"},
-        )
-
-        data = original.model_dump()
-        restored = Ether.model_validate(data)
-
-        assert restored.kind == original.kind
-        assert restored.schema_version == original.schema_version
-        assert restored.payload == original.payload
-        assert restored.metadata == original.metadata
-        assert restored.extra_fields == original.extra_fields
-
-    def test_json_serialization_round_trip(self) -> None:
-        """Test JSON serialization and deserialization."""
-        ether = Ether(
-            kind="tokens",
-            payload={"ids": [1, 2, 3]},
-            metadata={"vocab": "bert-base"},
-        )
-
-        json_str = ether.model_dump_json()
-        data = json.loads(json_str)
-        restored = Ether.model_validate(data)
-
-        assert restored.kind == ether.kind
-        assert restored.schema_version == ether.schema_version
-        assert restored.payload == ether.payload
-        assert restored.metadata == ether.metadata
-
-    def test_json_serialization_with_attachments(self) -> None:
-        """Test JSON serialization with attachments."""
-        attachment = Attachment(
-            id="emb-0",
-            inline_bytes=b"\x00\x00\x80\x3f",  # 1.0 in float32
-            media_type="application/x-raw-tensor",
-            codec="RAW_F32",
-            shape=[1],
-            dtype="float32",
-        )
-
-        ether = Ether(
-            kind="embedding",
-            payload={"dim": 1},
-            metadata={"source": "test"},
-            attachments=[attachment],
-        )
-
-        json_str = ether.model_dump_json()
-        data = json.loads(json_str)
-        restored = Ether.model_validate(data)
-
-        assert len(restored.attachments) == 1
-        assert restored.attachments[0].id == "emb-0"
-        assert restored.attachments[0].inline_bytes == b"\x00\x00\x80\x3f"
-
-
-class TestEtherSummary:
-    """Test Ether summary functionality."""
-
-    def test_summary_basic(self) -> None:
-        """Test basic summary functionality."""
-        ether = Ether(
-            kind="embedding",
-            schema_version=1,
-            payload={"dim": 768},
-            metadata={"source": "bert-base"},
-        )
-
-        summary = ether.summary()
-
-        assert summary["kind"] == "embedding"
-        assert summary["schema_version"] == 1
-        assert summary["payload_keys"] == ["dim"]
-        assert summary["metadata_keys"] == ["source"]
-        assert summary["extra_keys"] == []
-        assert summary["attachments"] == []
-        assert summary["source_model"] is None
-
-    def test_summary_with_nested_payload_and_metadata(self) -> None:
-        """Test summary with nested payload and metadata."""
-        ether = Ether(
-            kind="embedding",
-            payload={"vec": {"values": [1.0, 2.0], "dim": 2}},
-            metadata={"model": {"name": "bert", "version": "1.0"}},
-        )
-
-        summary = ether.summary()
-
-        assert summary["payload_keys"] == ["vec.dim", "vec.values"]
-        assert summary["metadata_keys"] == ["model.name", "model.version"]
-
-    def test_summary_with_attachments(self) -> None:
-        """Test summary with attachments."""
-        attachment1 = Attachment(
-            id="emb-0",
-            uri="shm://embeddings/12345",
-            media_type="application/x-raw-tensor",
-            codec="RAW_F32",
-        )
-        attachment2 = Attachment(
-            id="emb-1",
-            uri="shm://embeddings/12346",
-            media_type="application/x-raw-tensor",
-            codec="RAW_F32",
-        )
-
-        ether = Ether(
-            kind="embedding",
-            payload={},
-            metadata={},
-            attachments=[attachment1, attachment2],
-        )
-
-        summary = ether.summary()
-
-        assert summary["attachments"] == ["emb-0", "emb-1"]
-
-    def test_summary_with_extra_fields(self) -> None:
-        """Test summary with extra fields."""
-        ether = Ether(
-            kind="text",
-            payload={"text": "Hello"},
-            metadata={},
-            extra_fields={"note": "test", "priority": "high"},
-        )
-
-        summary = ether.summary()
-
-        assert summary["extra_keys"] == ["note", "priority"]
-
-
-class TestEtherEdgeCases:
-    """Test Ether edge cases and special scenarios."""
-
-    def test_ether_with_large_payload(self) -> None:
-        """Test Ether with large payload."""
-        large_payload = {"data": list(range(1000))}
-        ether = Ether(
-            kind="data",
-            payload=large_payload,
-            metadata={},
-        )
-
-        assert len(ether.payload["data"]) == 1000
-        assert ether.payload["data"][0] == 0
-        assert ether.payload["data"][-1] == 999
-
-    def test_ether_with_complex_metadata(self) -> None:
-        """Test Ether with complex metadata structure."""
-        complex_metadata = {
-            "trace_id": "123e4567-e89b-12d3-a456-426614174000",
-            "span_id": "456e7890-e89b-12d3-a456-426614174001",
-            "created_at": "2023-01-01T00:00:00Z",
-            "producer": "test-node",
-            "lineage": [{"node": "input", "version": "1.0", "ts": "2023-01-01T00:00:00Z"}],
-        }
-
-        ether = Ether(
-            kind="embedding",
-            payload={},
-            metadata=complex_metadata,
-        )
-
-        assert ether.metadata["trace_id"] == "123e4567-e89b-12d3-a456-426614174000"
-        assert len(ether.metadata["lineage"]) == 1
-
-    def test_ether_with_empty_attachments_list(self) -> None:
-        """Test Ether with empty attachments list."""
-        ether = Ether(
-            kind="text",
-            payload={"text": "Hello"},
-            metadata={},
-            attachments=[],
-        )
-
-        assert ether.attachments == []
-        assert len(ether.attachments) == 0
-
-    def test_ether_with_none_values_in_optional_fields(self) -> None:
-        """Test Ether with None values in optional fields."""
-        ether = Ether(
-            kind="embedding",
-            payload={"dim": 768},
-            metadata={"source": "bert-base"},
-            extra_fields={},
-            attachments=[],
-        )
-
-        assert ether.extra_fields == {}
-        assert ether.attachments == []
-
-
-class TestEtherImport:
-    """Test Ether import functionality."""
-
-    def test_import_from_package(self) -> None:
-        """Test that Ether can be imported from the package."""
-        from ether import Ether
-
-        ether = Ether(
-            kind="test",
-            payload={},
-            metadata={},
-        )
-
-        assert ether.kind == "test"
-
-    def test_ether_is_available_in_all(self) -> None:
-        """Test that Ether is available in __all__."""
-        from ether import __all__
-
-        assert "Ether" in __all__
-
-
-class TestEtherRegistration:
-    """Test Ether registration decorator functionality."""
-
-    def test_register_toy_model_success(self) -> None:
-        """Test successful registration of a toy model."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["embedding"],
-            metadata=["source", "dim"],
-            extra_fields="keep",
-            renames={"embedding": "vec.values", "dim": "vec.dim"},
-            kind="embedding",
-        )
-        class FooModel(BaseModel):
-            embedding: list[float]
-            source: str
-            dim: int
-            note: str = "extra"
-
-        # Verify registration was successful
-        assert len(_spec_registry) == 1
-        assert FooModel in _spec_registry
-
-        # Verify EtherSpec was created correctly
-        spec = _spec_registry[FooModel]
-        assert spec.payload_fields == ("embedding",)
-        assert spec.metadata_fields == ("source", "dim")
-        assert spec.extra_fields == "keep"
-        assert spec.renames == {"embedding": "vec.values", "dim": "vec.dim"}
-        assert spec.kind == "embedding"
-
-    def test_register_unknown_field_raises_error(self) -> None:
-        """Test that registering with unknown fields raises RegistrationError."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        with pytest.raises(RegistrationError, match="FooModel: unknown field 'unknown_field'"):
-
-            @Ether.register(
-                payload=["embedding"],
-                metadata=["source", "unknown_field"],  # This field doesn't exist
-                kind="embedding",
-            )
-            class FooModel(BaseModel):
-                embedding: list[float]
-                source: str
-                dim: int
-
-    def test_register_without_renames(self) -> None:
-        """Test registration without field renames."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["embedding"],
-            metadata=["source"],
-            extra_fields="ignore",
-        )
-        class BarModel(BaseModel):
-            embedding: list[float]
-            source: str
-
-        # Verify registration
-        assert len(_spec_registry) == 1
-        assert BarModel in _spec_registry
-
-        spec = _spec_registry[BarModel]
-        assert spec.payload_fields == ("embedding",)
-        assert spec.metadata_fields == ("source",)
-        assert spec.extra_fields == "ignore"
-        assert spec.renames == {}
-        assert spec.kind is None
-
-    def test_register_without_kind(self) -> None:
-        """Test registration without specifying kind."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["data"],
-            metadata=["info"],
-        )
-        class BazModel(BaseModel):
-            data: str
-            info: str
-
-        # Verify registration
-        assert len(_spec_registry) == 1
-        assert BazModel in _spec_registry
-
-        spec = _spec_registry[BazModel]
-        assert spec.kind is None
-
-    def test_register_with_empty_sequences(self) -> None:
-        """Test registration with empty payload and metadata sequences."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=[],
-            metadata=[],
-            extra_fields="error",
-        )
-        class EmptyModel(BaseModel):
-            extra_field: str = "test"
-
-        # Verify registration
-        assert len(_spec_registry) == 1
-        assert EmptyModel in _spec_registry
-
-        spec = _spec_registry[EmptyModel]
-        assert spec.payload_fields == ()
-        assert spec.metadata_fields == ()
-        assert spec.extra_fields == "error"
-
-    def test_register_multiple_models(self) -> None:
-        """Test registering multiple models in the same registry."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["field1"], metadata=[], kind="type1")
-        class Model1(BaseModel):
-            field1: str
-
-        @Ether.register(payload=["field2"], metadata=[], kind="type2")
-        class Model2(BaseModel):
-            field2: str
-
-        # Verify both models are registered
-        assert len(_spec_registry) == 2
-        assert Model1 in _spec_registry
-        assert Model2 in _spec_registry
-
-        # Verify different specs
-        spec1 = _spec_registry[Model1]
-        spec2 = _spec_registry[Model2]
-        assert spec1.kind == "type1"
-        assert spec2.kind == "type2"
-        assert spec1.payload_fields == ("field1",)
-        assert spec2.payload_fields == ("field2",)
-
-    def test_register_with_complex_renames(self) -> None:
-        """Test registration with complex field renames."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["embedding"],
-            metadata=["source"],
-            renames={
-                "embedding": "vec.values",
-                "source": "model.source",
-            },
-            kind="embedding",
-        )
-        class ComplexModel(BaseModel):
-            embedding: list[float]
-            source: str
-
-        # Verify registration
-        assert len(_spec_registry) == 1
-        assert ComplexModel in _spec_registry
-
-        spec = _spec_registry[ComplexModel]
-        assert spec.renames == {
-            "embedding": "vec.values",
-            "source": "model.source",
-        }
-
-    def test_register_decorator_returns_model(self) -> None:
-        """Test that the register decorator returns the model class."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["field"], metadata=[])
-        class TestModel(BaseModel):
-            field: str
-
-        # Verify the decorator returned the model class
-        assert TestModel.__name__ == "TestModel"
-        assert issubclass(TestModel, BaseModel)
-        assert TestModel in _spec_registry
-
-    def test_issue_26_acceptance_criteria(self) -> None:
-        """Test the specific acceptance criteria from issue #26."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        # Acceptance criteria 1: Unit test registers a toy FooModel; registry length == 1
-        @Ether.register(
-            payload=["embedding"],
-            metadata=["source", "dim"],
-            extra_fields="keep",
-            renames={"embedding": "vec.values", "dim": "vec.dim"},
-            kind="embedding",
-        )
-        class FooModel(BaseModel):
-            embedding: list[float]
-            source: str
-            dim: int
-            note: str = "extra"
-
-        # Verify registry length == 1
-        assert len(_spec_registry) == 1
-        assert FooModel in _spec_registry
-
-        # Acceptance criteria 2: Attempt to re-register same model raises RegistrationError
-        # Note: The current implementation allows re-registration, but we can test
-        # that the registration works correctly and the spec is updated
-        @Ether.register(
-            payload=["embedding"],
-            metadata=["source"],
-            kind="embedding_v2",
-        )
-        class FooModelV2(BaseModel):
-            embedding: list[float]
-            source: str
-
-        # Verify that a different model can be registered
-        assert len(_spec_registry) == 2
-        assert FooModelV2 in _spec_registry
-
-        # Verify that the specs are different
-        spec1 = _spec_registry[FooModel]
-        spec2 = _spec_registry[FooModelV2]
-        assert spec1.kind == "embedding"
-        assert spec2.kind == "embedding_v2"
-        assert spec1.metadata_fields == ("source", "dim")
-        assert spec2.metadata_fields == ("source",)
-
-
-class TestEtherAdapter:
-    """Test Ether adapter registration functionality."""
-
-    def test_adapter_registration(self) -> None:
-        """Test registering an adapter function."""
-        # Clear registry for clean test
-        _adapter_registry.clear()
-
-        class SourceModel(BaseModel):
-            field1: str
-
-        class DestModel(BaseModel):
-            field2: str
-
-        @Ether.adapter(SourceModel, DestModel)
-        def source_to_dest(eth: Ether) -> dict:
-            return {"field2": eth.payload.get("field1", "default")}
-
-        # Verify adapter was registered
-        assert len(_adapter_registry) == 1
-        assert (SourceModel, DestModel) in _adapter_registry
-        assert _adapter_registry[(SourceModel, DestModel)] == source_to_dest
-
-    def test_adapter_decorator_returns_function(self) -> None:
-        """Test that the adapter decorator returns the function."""
-        # Clear registry for clean test
-        _adapter_registry.clear()
-
-        class SourceModel(BaseModel):
-            field1: str
-
-        class DestModel(BaseModel):
-            field2: str
-
-        @Ether.adapter(SourceModel, DestModel)
-        def test_adapter(eth: Ether) -> dict:
-            return {"field2": "test"}
-
-        # Verify the decorator returned the function
-        assert test_adapter.__name__ == "test_adapter"
-        assert callable(test_adapter)
-
-
-class TestEtherFromModel:
-    """Test Ether.from_model() factory method."""
-
-    def test_from_model_basic_conversion(self) -> None:
-        """Test basic conversion from model to Ether."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["embedding", "dim"],
-            metadata=["source"],
-            extra_fields="keep",
-            renames={"embedding": "vec.values", "dim": "vec.dim"},
-            kind="embedding",
-        )
-        class FooModel(BaseModel):
-            embedding: list[float]
-            source: str
-            dim: int
-            note: str = "extra"
-
-        # Create model instance
-        model = FooModel(embedding=[1.0, 2.0, 3.0], source="bert", dim=3, note="test")
-
-        # Convert to Ether
-        ether = Ether.from_model(model)
-
-        # Verify conversion
-        assert ether.kind == "embedding"
-        assert ether.schema_version == 1
-        assert ether.payload == {"vec": {"values": [1.0, 2.0, 3.0], "dim": 3}}
-        assert ether.metadata == {"source": "bert"}
-        assert ether.extra_fields == {"note": "test"}
-        assert ether._source_model == FooModel
-
-    def test_from_model_with_custom_schema_version(self) -> None:
-        """Test from_model with custom schema version."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["data"], metadata=[], kind="test")
-        class TestModel(BaseModel):
-            data: str
-
-        model = TestModel(data="test")
-        ether = Ether.from_model(model, schema_version=2)
-
-        assert ether.schema_version == 2
-        assert ether.kind == "test"
-
-    def test_from_model_unregistered_model_raises_error(self) -> None:
-        """Test that from_model raises error for unregistered model."""
-
-        class UnregisteredModel(BaseModel):
-            field: str
-
-        model = UnregisteredModel(field="test")
-
-        with pytest.raises(RegistrationError, match="UnregisteredModel not registered"):
-            Ether.from_model(model)
-
-    def test_from_model_extra_fields_error_policy(self) -> None:
-        """Test from_model with extra_fields='error' policy."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["data"],
-            metadata=[],
-            extra_fields="error",
-            kind="test",
-        )
-        class TestModel(BaseModel):
-            data: str
-            extra_field: str = "should_fail"
-
-        model = TestModel(data="test", extra_field="fail")
-
-        with pytest.raises(ConversionError, match="Extra fields not allowed"):
-            Ether.from_model(model)
-
-    def test_from_model_extra_fields_keep_policy(self) -> None:
-        """Test from_model with extra_fields='keep' policy."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["data"],
-            metadata=[],
-            extra_fields="keep",
-            kind="test",
-        )
-        class TestModel(BaseModel):
-            data: str
-            extra_field: str = "should_keep"
-
-        model = TestModel(data="test", extra_field="keep_me")
-
-        ether = Ether.from_model(model)
-
-        assert ether.extra_fields == {"extra_field": "keep_me"}
-
-    def test_from_model_extra_fields_ignore_policy(self) -> None:
-        """Test from_model with extra_fields='ignore' policy."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["data"],
-            metadata=[],
-            extra_fields="ignore",
-            kind="test",
-        )
-        class TestModel(BaseModel):
-            data: str
-            extra_field: str = "should_ignore"
-
-        model = TestModel(data="test", extra_field="ignore_me")
-
-        ether = Ether.from_model(model)
-
-        assert ether.extra_fields == {}
-
-    def test_from_model_with_nested_renames(self) -> None:
-        """Test from_model with nested field renames."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["embedding"],
-            metadata=["source"],
-            renames={
-                "embedding": "vec.values",
-                "source": "model.source",
-            },
-            kind="embedding",
-        )
-        class TestModel(BaseModel):
-            embedding: list[float]
-            source: str
-
-        model = TestModel(embedding=[1.0, 2.0], source="bert")
-
-        ether = Ether.from_model(model)
-
-        assert ether.payload == {"vec": {"values": [1.0, 2.0]}}
-        assert ether.metadata == {"model": {"source": "bert"}}
-
-    def test_from_model_without_kind(self) -> None:
-        """Test from_model with model registered without kind."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["data"], metadata=[])
-        class TestModel(BaseModel):
-            data: str
-
-        model = TestModel(data="test")
-
-        ether = Ether.from_model(model)
-
-        assert ether.kind == ""  # Empty string when no kind specified
-
-
-class TestEtherAsModel:
-    """Test Ether.as_model() conversion method."""
-
-    def test_as_model_basic_conversion(self) -> None:
-        """Test basic conversion from Ether to model."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["embedding"],
-            metadata=["source"],
-            kind="embedding",
-        )
-        class TargetModel(BaseModel):
-            embedding: list[float]
-            source: str
-
-        # Create Ether
-        ether = Ether(
-            kind="embedding",
-            payload={"embedding": [1.0, 2.0, 3.0]},
-            metadata={"source": "bert"},
-        )
-
-        # Convert to model
-        model = ether.as_model(TargetModel)
-
-        # Verify conversion
-        assert model.embedding == [1.0, 2.0, 3.0]
-        assert model.source == "bert"
-
-    def test_as_model_with_renames(self) -> None:
-        """Test as_model with field renames."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["embedding"],
-            metadata=["source"],
-            renames={
-                "embedding": "vec.values",
-                "source": "model.source",
-            },
-            kind="embedding",
-        )
-        class TargetModel(BaseModel):
-            embedding: list[float]
-            source: str
-
-        # Create Ether with nested structure
-        ether = Ether(
-            kind="embedding",
-            payload={"vec": {"values": [1.0, 2.0]}},
-            metadata={"model": {"source": "bert"}},
-        )
-
-        # Convert to model
-        model = ether.as_model(TargetModel)
-
-        # Verify conversion
-        assert model.embedding == [1.0, 2.0]
-        assert model.source == "bert"
-
-    def test_as_model_unregistered_model_raises_error(self) -> None:
-        """Test that as_model raises error for unregistered model."""
-
-        class UnregisteredModel(BaseModel):
-            field: str
-
-        ether = Ether(kind="test", payload={}, metadata={})
-
-        with pytest.raises(RegistrationError, match="UnregisteredModel not registered"):
-            ether.as_model(UnregisteredModel)
-
-    def test_as_model_missing_required_fields_raises_error(self) -> None:
-        """Test that as_model raises error for missing required fields."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["embedding"], metadata=["source"], kind="embedding")
-        class TargetModel(BaseModel):
-            embedding: list[float]
-            source: str
-            required_field: str  # This field is required but not provided
-
-        ether = Ether(
-            kind="embedding",
-            payload={"embedding": [1.0, 2.0]},
-            metadata={"source": "bert"},
-        )
-
-        with pytest.raises(ConversionError, match="Missing required fields"):
-            ether.as_model(TargetModel)
-
-    def test_as_model_require_kind_mismatch_raises_error(self) -> None:
-        """Test that as_model with require_kind=True raises error for kind mismatch."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["data"], metadata=[], kind="type1")
-        class TargetModel(BaseModel):
-            data: str
-
-        ether = Ether(
-            kind="type2",  # Different kind
-            payload={"data": "test"},
-            metadata={},
-        )
-
-        with pytest.raises(ConversionError, match="Kind mismatch"):
-            ether.as_model(TargetModel, require_kind=True)
-
-    def test_as_model_require_kind_success(self) -> None:
-        """Test that as_model with require_kind=True succeeds for matching kinds."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["data"], metadata=[], kind="test")
-        class TargetModel(BaseModel):
-            data: str
-
-        ether = Ether(
-            kind="test",  # Matching kind
-            payload={"data": "test"},
-            metadata={},
-        )
-
-        model = ether.as_model(TargetModel, require_kind=True)
-
-        assert model.data == "test"
-
-    def test_as_model_with_adapter(self) -> None:
-        """Test as_model using adapter function."""
-        # Clear registries for clean test
-        _spec_registry.clear()
-        _adapter_registry.clear()
-
-        @Ether.register(payload=["embedding"], metadata=["source"], kind="embedding")
-        class SourceModel(BaseModel):
-            embedding: list[float]
-            source: str
-
-        @Ether.register(payload=["count"], metadata=[], kind="count")
-        class TargetModel(BaseModel):
-            count: int
-
-        @Ether.adapter(SourceModel, TargetModel)
-        def source_to_target(eth: Ether) -> dict:
-            return {"count": len(eth.payload["embedding"])}
-
-        # Create Ether from source model
-        source_model = SourceModel(embedding=[1.0, 2.0, 3.0], source="bert")
-        ether = Ether.from_model(source_model)
-
-        # Convert to target model using adapter
-        target_model = ether.as_model(TargetModel)
-
-        # Verify conversion
-        assert target_model.count == 3
-
-    def test_as_model_from_extra_fields(self) -> None:
-        """Test as_model picking fields from extra_fields."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["data"], metadata=[], kind="test")
-        class TargetModel(BaseModel):
-            data: str
-            extra_field: str
-
-        ether = Ether(
-            kind="test",
-            payload={"data": "test"},
-            metadata={},
-            extra_fields={"extra_field": "extra_value"},
-        )
-
-        model = ether.as_model(TargetModel)
-
-        assert model.data == "test"
-        assert model.extra_field == "extra_value"
-
-    def test_as_model_round_trip(self) -> None:
-        """Test round-trip conversion: Model -> Ether -> Model."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["embedding", "dim"],
-            metadata=["source"],
-            extra_fields="keep",
-            renames={"embedding": "vec.values", "dim": "vec.dim"},
-            kind="embedding",
-        )
-        class FooModel(BaseModel):
-            embedding: list[float]
-            source: str
-            dim: int
-            note: str = "extra"
-
-        # Original model
-        original = FooModel(embedding=[1.0, 2.0, 3.0], source="bert", dim=3, note="test")
-
-        # Convert to Ether
-        ether = Ether.from_model(original)
-
-        # Convert back to model
-        restored = ether.as_model(FooModel)
-
-        # Verify round-trip
-        assert restored.embedding == original.embedding
-        assert restored.source == original.source
-        assert restored.dim == original.dim
-        assert restored.note == original.note
-
-    def test_as_model_pick_from_extras_by_renamed_key(self) -> None:
-        """Test as_model picking fields from extras using renamed keys."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["data"],
-            metadata=[],
-            renames={"data": "payload.data"},
-            kind="test",
-        )
-        class TargetModel(BaseModel):
-            data: str
-
-        # Create Ether with renamed key in extras
-        ether = Ether(
-            kind="test",
-            payload={},
-            metadata={},
-            extra_fields={"payload.data": "test_value"},
-        )
-
-        model = ether.as_model(TargetModel)
-
-        assert model.data == "test_value"
-
-    def test_as_model_pick_from_extras_by_model_field_name(self) -> None:
-        """Test as_model picking fields from extras using model field name."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["data"],
-            metadata=[],
-            renames={"data": "payload.data"},
-            kind="test",
-        )
-        class TargetModel(BaseModel):
-            data: str
-
-        # Create Ether with model field name in extras
-        ether = Ether(
-            kind="test",
-            payload={},
-            metadata={},
-            extra_fields={"data": "test_value"},
-        )
-
-        model = ether.as_model(TargetModel)
-
-        assert model.data == "test_value"
-
-    def test_as_model_validation_error_without_missing_fields(self) -> None:
-        """Test as_model when ValidationError occurs but no fields are missing."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["data"], metadata=[], kind="test")
-        class TargetModel(BaseModel):
-            data: str
-
-        # Create Ether with invalid data type
-        ether = Ether(
-            kind="test",
-            payload={"data": 123},  # Should be str, not int
-            metadata={},
-        )
-
-        # This should raise ValidationError but not ConversionError
-        with pytest.raises(ValidationError):
-            ether.as_model(TargetModel)
-
-    def test_as_model_pick_returns_false_when_field_not_found(self) -> None:
-        """Test as_model when pick function returns False, None for missing fields."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["data"], metadata=[], kind="test")
-        class TargetModel(BaseModel):
-            data: str
-            missing_field: str  # This field is not in payload, metadata, or extras
-
-        # Create Ether without the required field
-        ether = Ether(
-            kind="test",
-            payload={"data": "test"},
-            metadata={},
-            extra_fields={},
-        )
-
-        # This should raise ConversionError for missing required field
-        with pytest.raises(ConversionError, match="Missing required fields"):
-            ether.as_model(TargetModel)
-
-    def test_as_model_pick_returns_false_for_optional_field(self) -> None:
-        """Test as_model when pick function returns False, None for optional fields."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["data"], metadata=[], kind="test")
-        class TargetModel(BaseModel):
-            data: str
-            optional_field: str | None = None  # Optional field not in payload/metadata/extras
-
-        # Create Ether without the optional field
-        ether = Ether(
-            kind="test",
-            payload={"data": "test"},
-            metadata={},
-            extra_fields={},
-        )
-
-        # This should succeed and use the default value
-        model = ether.as_model(TargetModel)
-        assert model.data == "test"
-        assert model.optional_field is None
-
-    def test_as_model_pick_returns_false_for_renamed_field_not_in_sources(self) -> None:
-        """Test as_model when pick function returns False, None for renamed field not in any source."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["data"], metadata=[], renames={"data": "payload.data", "missing": "payload.missing"}, kind="test"
-        )
-        class TargetModel(BaseModel):
-            data: str
-            missing: str | None = None  # Optional field with rename, not in any source
-
-        # Create Ether with data but without the missing field
-        ether = Ether(
-            kind="test",
-            payload={"payload": {"data": "test"}},  # Only data, no missing
-            metadata={},
-            extra_fields={},
-        )
-
-        # This should succeed and use the default value for missing
-        model = ether.as_model(TargetModel)
-        assert model.data == "test"
-        assert model.missing is None
-
-    def test_as_model_pick_returns_false_for_metadata_field_not_found(self) -> None:
-        """Test as_model when pick function returns False, None for metadata field not found."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(
-            payload=["data"], metadata=["metadata_field"], renames={"metadata_field": "meta.field"}, kind="test"
-        )
-        class TargetModel(BaseModel):
-            data: str
-            metadata_field: str | None = None  # Optional metadata field with rename
-
-        # Create Ether without the metadata field
-        ether = Ether(
-            kind="test",
-            payload={"data": "test"},
-            metadata={},  # No metadata
-            extra_fields={},
-        )
-
-        # This should succeed and use the default value
-        model = ether.as_model(TargetModel)
-        assert model.data == "test"
-        assert model.metadata_field is None
-
-
-class TestEtherConstructorWithModel:
-    """Test Ether constructor with model argument."""
-
-    def test_ether_constructor_with_model(self) -> None:
-        """Test Ether constructor with model argument."""
-        # Clear registry for clean test
-        _spec_registry.clear()
-
-        @Ether.register(payload=["data"], metadata=[], kind="test")
-        class TestModel(BaseModel):
-            data: str
-
-        model = TestModel(data="test")
-
-        # Use constructor with model
-        ether = Ether(model)
-
-        # Verify conversion
-        assert ether.kind == "test"
-        assert ether.payload == {"data": "test"}
-        assert ether._source_model == TestModel
-
-    def test_ether_constructor_with_model_and_kwargs_raises_error(self) -> None:
-        """Test that Ether constructor with model and kwargs raises error."""
-
-        class TestModel(BaseModel):
-            data: str
-
-        model = TestModel(data="test")
-
-        with pytest.raises(TypeError):
-            Ether(model, kind="test")  # Should not accept both model and kwargs
